+++ src/OpenTelemetry/Trace/TracerProviderBuilderSdk.cs
@@ -32,7 +32,6 @@ namespace OpenTelemetry.Trace
         private readonly List<BaseProcessor<Activity>> processors = new List<BaseProcessor<Activity>>();
         private readonly List<string> sources = new List<string>();
         private ResourceBuilder resourceBuilder = ResourceBuilder.CreateDefault();
-        private Action<ResourceBuilder> configureResource;
         private Sampler sampler = new ParentBasedSampler(new AlwaysOnSampler());
 
         internal TracerProviderBuilderSdk()
@@ -97,39 +96,18 @@ namespace OpenTelemetry.Trace
         /// <returns>Returns <see cref="TracerProviderBuilder"/> for chaining.</returns>
         internal TracerProviderBuilder SetSampler(Sampler sampler)
         {
-            if (sampler == null)
-            {
-                throw new ArgumentNullException(nameof(sampler));
-            }
-
-            this.sampler = sampler;
+            this.sampler = sampler ?? throw new ArgumentNullException(nameof(sampler));
             return this;
         }
 
         /// <summary>
         /// Sets the <see cref="Resource"/> describing the app associated with all traces. Overwrites currently set resource.
         /// </summary>
-        /// <param name="resource">Resource to be associate with all traces.</param>
+        /// <param name="resourceBuilder"><see cref="ResourceBuilder"/> to be associate with all traces.</param>
         /// <returns>Returns <see cref="TracerProviderBuilder"/> for chaining.</returns>
-        internal TracerProviderBuilder SetResource(Resource resource)
+        internal TracerProviderBuilder SetResourceBuilder(ResourceBuilder resourceBuilder)
         {
-            if (resource == null)
-            {
-                throw new ArgumentNullException(nameof(resource));
-            }
-
-            this.resourceBuilder = new ResourceBuilder().AddResource(resource);
-            return this;
-        }
-
-        /// <summary>
-        /// Registers a callback that will be executed to configure the <see cref="Resource"/> for the <see cref="TracerProvider"/>.
-        /// </summary>
-        /// <param name="configureResource"><see cref="ResourceBuilder"/>.</param>
-        /// <returns>Returns <see cref="TracerProviderBuilder"/> for chaining.</returns>
-        internal TracerProviderBuilder ConfigureResource(Action<ResourceBuilder> configureResource)
-        {
-            this.configureResource = configureResource ?? throw new ArgumentNullException(nameof(configureResource));
+            this.resourceBuilder = resourceBuilder ?? throw new ArgumentNullException(nameof(resourceBuilder));
             return this;
         }
 
@@ -152,8 +130,6 @@ namespace OpenTelemetry.Trace
 
         internal TracerProvider Build()
         {
-            this.configureResource?.Invoke(this.resourceBuilder);
-
             return new TracerProviderSdk(
                 this.resourceBuilder.Build(),
                 this.sources,