+++ test/OpenTelemetry.Tests/Resources/ResourceBuilderTests.cs
@@ -26,7 +26,7 @@ namespace OpenTelemetry.Resources.Tests
         [Fact]
         public void ServiceResource_ServiceName()
         {
-            var resource = new ResourceBuilder().AddService("my-service").Build();
+            var resource = ResourceBuilder.CreateEmpty().AddService("my-service").Build();
             Assert.Equal(2, resource.Attributes.Count());
             Assert.Contains(new KeyValuePair<string, object>(ResourceSemanticConventions.AttributeServiceName, "my-service"), resource.Attributes);
             Assert.Single(resource.Attributes.Where(kvp => kvp.Key == ResourceSemanticConventions.AttributeServiceName));
@@ -36,7 +36,7 @@ namespace OpenTelemetry.Resources.Tests
         [Fact]
         public void ServiceResource_ServiceNameAndInstance()
         {
-            var resource = new ResourceBuilder().AddService("my-service", serviceInstanceId: "123").Build();
+            var resource = ResourceBuilder.CreateEmpty().AddService("my-service", serviceInstanceId: "123").Build();
             Assert.Equal(2, resource.Attributes.Count());
             Assert.Contains(new KeyValuePair<string, object>(ResourceSemanticConventions.AttributeServiceName, "my-service"), resource.Attributes);
             Assert.Contains(new KeyValuePair<string, object>(ResourceSemanticConventions.AttributeServiceInstance, "123"), resource.Attributes);
@@ -45,7 +45,7 @@ namespace OpenTelemetry.Resources.Tests
         [Fact]
         public void ServiceResource_ServiceNameAndInstanceAndNamespace()
         {
-            var resource = new ResourceBuilder().AddService("my-service", "my-namespace", serviceInstanceId: "123").Build();
+            var resource = ResourceBuilder.CreateEmpty().AddService("my-service", "my-namespace", serviceInstanceId: "123").Build();
             Assert.Equal(3, resource.Attributes.Count());
             Assert.Contains(new KeyValuePair<string, object>(ResourceSemanticConventions.AttributeServiceName, "my-service"), resource.Attributes);
             Assert.Contains(new KeyValuePair<string, object>(ResourceSemanticConventions.AttributeServiceInstance, "123"), resource.Attributes);
@@ -55,7 +55,7 @@ namespace OpenTelemetry.Resources.Tests
         [Fact]
         public void ServiceResource_ServiceNameAndInstanceAndNamespaceAndVersion()
         {
-            var resource = new ResourceBuilder().AddService("my-service", "my-namespace", "semVer:1.2.3", serviceInstanceId: "123").Build();
+            var resource = ResourceBuilder.CreateEmpty().AddService("my-service", "my-namespace", "semVer:1.2.3", serviceInstanceId: "123").Build();
             Assert.Equal(4, resource.Attributes.Count());
             Assert.Contains(new KeyValuePair<string, object>(ResourceSemanticConventions.AttributeServiceName, "my-service"), resource.Attributes);
             Assert.Contains(new KeyValuePair<string, object>(ResourceSemanticConventions.AttributeServiceInstance, "123"), resource.Attributes);
@@ -66,7 +66,7 @@ namespace OpenTelemetry.Resources.Tests
         [Fact]
         public void ServiceResource_AutoGenerateServiceInstanceIdOff()
         {
-            var resource = new ResourceBuilder().AddService("my-service", autoGenerateServiceInstanceId: false).Build();
+            var resource = ResourceBuilder.CreateEmpty().AddService("my-service", autoGenerateServiceInstanceId: false).Build();
             Assert.Single(resource.Attributes);
             Assert.Contains(new KeyValuePair<string, object>(ResourceSemanticConventions.AttributeServiceName, "my-service"), resource.Attributes);
         }
@@ -74,7 +74,7 @@ namespace OpenTelemetry.Resources.Tests
         [Fact]
         public void ClearTest()
         {
-            var resource = new ResourceBuilder()
+            var resource = ResourceBuilder.CreateEmpty()
                 .AddTelemetrySdk()
                 .Clear()
                 .AddService("my-service", autoGenerateServiceInstanceId: false)